chatbot/requirements.txt
Flask==3.0.0
cohere==4.37
python-dotenv==1.0.0
requests
feedparser

